added all basic homepage components and organized file structure
