+++ src/components/layout/layout.js
@@ -9,7 +9,7 @@ import React from "react"
 import PropTypes from "prop-types"
 import { useStaticQuery, graphql } from "gatsby"
 
-import Header from "./header"
+import Header from "../header"
 import "./layout.css"
 
 const Layout = ({ children }) => {
