+++ src/components/what-we-do/industry-bg/industry-bg.js
