+++ src/pages/what-we-do.js
@@ -1,11 +1,11 @@
 import React from "react"
 import { graphql } from "gatsby"
 import get from "lodash/get"
-import Layout from "../components/layout"
+import Layout from "../components/layout/layout"
 import SEO from "../components/seo"
-import History from "../components/what-we-do/history"
-import IndustryBG from "../components/what-we-do/industry-bg"
-import ProgrammingCard from "../components/what-we-do/programming-card"
+import History from "../components/what-we-do/history/history"
+import IndustryBG from "../components/what-we-do/industry-bg/industry-bg"
+import ProgrammingCard from "../components/what-we-do/programming-cards/programming-card"
 
 const WhatWeDo = props => {
   const whatWeDo = get(props, "data.contentfulWhatWeDo")
